+++ src/Blazor.Server.UI/Components/Localization/LanguageSelector.razor
@@ -1,6 +1,7 @@
 @using System.Globalization
 @using Microsoft.Extensions.Options
 @using Blazor.Server.UI.Services
+@using Blazor.Server.UI.Services.Layout
 <MudTooltip Text="@CurrentLanguage">
     <MudMenu Class="mx-1" Icon="@Icons.Material.Outlined.Translate" Color="Color.Default" Dense="true">
         @if (SupportedLanguages is not null)
