+++ src/Blazor.Server.UI/Components/Shared/SideMenu.razor.cs
@@ -1,5 +1,6 @@
 using Blazor.Server.UI.Models.SideMenu;
 using Blazor.Server.UI.Services;
+using Blazor.Server.UI.Services.Layout;
 using Blazor.Server.UI.Services.Navigation;
 using CleanArchitecture.Blazor.Application.Features.Fluxor;
 using Fluxor;
