+++ src/Blazor.Server.UI/Components/Shared/Themes/ThemesButton.razor.cs
@@ -1,4 +1,5 @@
 using Blazor.Server.UI.Services;
+using Blazor.Server.UI.Services.Layout;
 using Microsoft.AspNetCore.Components.Web;
 
 namespace Blazor.Server.UI.Components.Shared.Themes;
