+++ src/Blazor.Server.UI/Pages/SystemManagement/Components/LogsLineCharts.razor
@@ -1,6 +1,7 @@
 ﻿@using CleanArchitecture.Blazor.Application.Features.Loggers.DTOs
 @using ApexCharts
 @using Blazor.Server.UI.Services
+@using Blazor.Server.UI.Services.Layout
 @implements IDisposable
 @inject IStringLocalizer<Logs> L
